+++ lib/src/bootstrap_selectable_text_parameters.dart
@@ -61,6 +61,12 @@ class BSSelectParams {
   List<Shadow>? shadows;
   TextBaseline? textBaseline;
   double? wordSpacing;
+  double Function(
+      BuildContext context, // Build Context used to build text widget
+      String textType, { // p, or h1-h6
+      required double small, // default BSBreakPoints value unless textSizeSmall is passed
+      required double large, // default BSBreakPoints value unless textSizeLarge is passed
+  })? customBreakpoints;
 
   BSSelectParams(
     this.text, {
@@ -116,5 +122,6 @@ class BSSelectParams {
     this.shadows,
     this.textBaseline,
     this.wordSpacing,
+    this.customBreakpoints,
   });
 }