+++ lib/src/bootstrap_text_parameters.dart
@@ -41,6 +41,12 @@ class BSTextParams {
   double? textScaleFactor;
   TextWidthBasis? textWidthBasis;
   Key? textKey;
+  double Function(
+    BuildContext context, // Build Context used to build text widget
+    String textType, { // p, or h1-h6
+    required double small, // default BSBreakPoints value unless textSizeSmall is passed
+    required double large, // default BSBreakPoints value unless textSizeLarge is passed
+  })? customBreakpoints;
 
   /// text is the only required param
   BSTextParams(
@@ -81,5 +87,6 @@ class BSTextParams {
     this.textScaleFactor,
     this.textWidthBasis,
     this.textKey,
+    this.customBreakpoints,
   });
 }